+++ kopy-gradle-plugin/main/kotlin/com/javiersc/kotlin/kopy/gradle/plugin/KopyExtension.kt
@@ -1,5 +1,6 @@
 package com.javiersc.kotlin.kopy.gradle.plugin
 
+import com.javiersc.kotlin.kopy.args.KopyFunctions
 import com.javiersc.kotlin.kopy.args.KopyVisibility
 import javax.inject.Inject
 import org.gradle.api.model.ObjectFactory
@@ -12,6 +13,9 @@ constructor(
     private val providers: ProviderFactory,
     objects: ObjectFactory,
 ) {
+
+    public val functions: Property<KopyFunctions> =
+        objects.property(KopyFunctions::class.java).convention(KopyFunctions.All)
 
     public val visibility: Property<KopyVisibility> =
         objects.property(KopyVisibility::class.java).convention(KopyVisibility.Auto)